Add files via upload
India's economy is expanding quickly on a global scale, and during the course of the next ten years, it is anticipated that India will rank among the top 3 economies on the planet. In contrast to a downturn in Indian corporate earnings brought on by excess capacity and banks' incapacity to extend credit, the stock market, specifically the Bombay Stock Exchange, has done well. The purpose of the current study is to examine any relationships, if any, between the BSE Sensex index and macroeconomic indicators such as the Foreign direct investment, Real gross domestic product, Inflation rate and Interest Rate in India from the the time period between 1991and 2021. The investigation aims to establish link between the variables based on the tests of Johansen Cointegration, Granger Causality, and Vector error-correction mechanism, and to find the strength of the relationship among the parameters that are independent and the dependent parameter. The analysis using the Vector Error Correction Model (VECM) supports the existence of a long-run causal relationship between the macroeconomic variables of the Index of Foreign direct Investment, inflation, interest rates, Real gross domestic product and BSE Sensex.

Keywords: Sensex, FDI, Interest rate, Inflation rate, Augmented dickey fuller, Johansen Cointegration, Granger Causality, and the Vector Error Correction Model